+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Lists that still have elements can now be deleted.
+  (See: [!627](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/627))
+- Lists can now be updated with an empty element list.
+  (See: [!627](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/627))
 
 ## [0.37.0] - 2023-09-08
 ### Added